Better handling of errors in subscribers
Fixes #5

We were not handling subscriber errors aggressively enough.
The idea was that, because publish and subscribe are decoupled,
a failure in a subscriber should NOT be a critical error
in the app- it should be a failure of just that subscriber.
And we log the error out of convenience.

This makes sense in theory, in a truly distributed pub/sub system,
but in practice, clients depend on subscriptions to publish
properly; for example, if model events fail to publish
(because, say, they enqueue a job into Redis)
many critical components of a system will not work.

Instead, the default behavior should be to raise an error
if any subscriber fails. This would cause code that fails to publish
to raise an error on the `publish` call, returning a 500,
failing the job, etc.

To continue the old behavior,
you can use `Amigo.on_publish_error = proc {}`.

This also improves the logging so that we log the full event details,
rather than just the representation of the object.
This means that, if something critical fails to publish,
there is still a record of the event for future use in repairs.