Handle audit logger perform_async failure
Fixes #6

If the audit logger job cannot `perform_async`,
we would get an error in the subscriber.
This is not great, since it means we don't get the event audited.

Instead, if we cannot `.perform_async` we can `.new.perform`.
This ensures we still get the event audited,
even if Redis is down.