Disable backoff behavior when near OOM conditions
When Redis is using a lot of memory (90% of available),
we want to disable queue backoff behavior.
There is a significant risk that the backoff
behavior will take jobs from the queue,
and immediately try and reschedule them.
If that happens in an OOM condition,
the re-push will fail and the job can be lost.

Additionally, the backoff behavior causes delays that
slow down the clearing of the queue.

In these high-memory-utilization conditions,
it makes more sense to disable the backoff logic
and just brute force to try to get through the queue.