Rewrite autoscaler (checker and handler); add web latency checker (#20)
NOTE: This change is backwards incompatible with current Autoscaler usage. Please see Autoscaler docs for upgrade instructions, which should be straightforward (and results in a simpler application)

---

Autoscaling is now broken into a 'checker' and a 'handler'.

Checkers can check Sidekiq queue latency, or web request latency, whatever.

Handlers take an action, like logging, sending to Sentry, or scaling dynos in Heroku. There is also a chain handler.

It's now possible to easily create a web dyno autoscaler. It also cleaned up a lot of the internal complexity with the autoscaler parameters.

Also fixes coverage reporting in CI.