Add RetryOrDie callback support
Callers can provide a subclass with two methods that are looked for:

If on_retry is defined, it is called with (worker instance, job hash).
If on_retry returns +:skip+, do NOT retry (do not send to the retry set).

If on_die is defined, it is called with (worker instance, job hash).
If on_die returns +:skip+, do NOT send to the dead set.